requirements.txt
asgiref==3.2.7
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
